Spike for location already taken
+ Board#inspect does double duty and also outputs the list of available
locations for player's selection to be validated against. Validation
logic in Player feels awkward and misplaced because it's very tell don't ask.
+ Location selection prompt loops until valid input is obtained.
Interesting side effect of this is that it addresses the validation
concerns that were alluded to in the first story that obtained the
selected location from the player.